Build gcc before binutils.
Per Alex Crichton's request that we build gcc as early as possible.

The gcc build is the slowest step, so the earlier we do it, the more
likely it is that updates to other components will be able to use a
cached image with the gcc build.

Since binutils is a dependency of gcc, this could in theory change
the output. It is able to compile the rest of the components and
rust itself without error, though.