Build gcc first.
The gcc build step takes by far the longest, and we don't expect
to update it as often as later dependencies. So we trade the
improved security of using a newer-than-system openssl and curl
to download source for the convenience of faster rebuilds by making
it more likely the gcc build will be in docker's cache.