Use pip instead of easy_install.
I can no longer build this container because of SSL cert validation
errors running (but not downloading) ez_setup.py. This might be
related to turning on cert validation (!) in 2.7.9-ish per PEP 0476,
but it worked a month ago with python 2.7.10.

Perhaps it's something about the build config or system cert
locations which have changed. I wasn't able to find the issue.

In any case, switching to the more modern pip tool avoids the issue.
Python is supposed to include pip these days, but our build doesn't
provide it.