Fix cargo call to finish_dist
Use the `cargo_platforms` as the start for learning all hosts that Cargo has
nightlies for (e.g. if you have an auto platform then you have a dist platform
assumed). Also uniq-ify this set as there are repeats (e.g. the set of "all
hosts" for the "linux-32" builder is the same as the "linux-64" builder).
Finally, also don't set any nogate builders as we're gating on everything for
Cargo right now.