Make the system cert.pem available to python.
We build and install openssl in /rustroot/ but the system's
TLS certificate chain is in /etc/pki/tls. This prevents
python from validating the certificates on https urls,
blocking tooltool and probably also why easy-install
was failing before we switched to pip. I guess all this
used to work only because python wasn't validating certs
in earlier 2.7 releases.

This wasn't a blocker because most of the image scripts
use curl for download. I assume--but haven't verified--
that this works because curl relies on the /etc/ path,
overriding the openssl default.