Use an http-only mirror for curl source.
Recently haxx.se set a redirect to https on their http download
urls, but the centos 5 curl we're using is too old to form
secure connections. Fall back to a mirror. This is likely
brittle too, but we still verify the checksum so it should
be better than the current bit-rot.