linux-cross: add armv7-linux-gnu toolchain
When using rustbuild to build libstd/rustc for the armv7-unknown-linux-gnueabihf target we use this
toolchain instead of the arm-linux-gnueabihf one.

This commit also installs cmake and gives a home directory to the rustbuild user. Both are required
to be able to use rustbuild to cross compile rustc within this image.