Add custom toolchains for powerpc* and s390x
Using Ubuntu's cross-toolchains meant these targets were depending on
glibc symbols from Ubuntu 16.04.  And if that host is ever updated to a
new release, the toolchains would raise the bar too.

This switches powerpc* and s390x to use crosstool-ng toolchains, with
the kernel and glibc configured approximately like RHEL7.  That should
also be compatible with Debian 8 (jessie) and Ubuntu 14.04 (trusty).