Push powerpc* and s390x back even further.
For crosstool-ng with powerpc, powerpc64, and s390x, a few small patches
are now applied to glibc-2.12 to get it to compile with newer GCC.  This
should now be compatible with RHEL6, Debian 7 (wheezy), and Ubuntu 12.04
(precise).

For powerpc64le, the challenge was that only glibc-2.19 officially added
support, but el7 backported those changes to glibc-2.17.  The backport
patches are complex and numerous, so instead of trying to push those
into crosstool-ng, just use glibc binaries directly from centos7 and
build the toolchain manually.