Download gcc without tls.
The new certificate for ftp.gnu.org doesn't validate on centos:5
because it thinks the hostname is alpha.gnu.org, which isn't listed
on the cert. Work around by dropping TLS protection on the download,
relying on just the checksum for validation.

Once we've bootstrapped our own openssl and curl we can download
things properly, so another fix would be to move gcc later in the
build order.